README: add section about adding the "extractor" directory to the llvm-clang-tree
do it so that we don't have to copy the directory into the tree. cmake
itself does not want to include subdirectories which are outside of the
source tree... using a symlink to make it look "local" is more like a
hack ;-) well cmake is dumb enough to eat this. as long as it works we
are happy with it.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>